update treearray lib - supports tree mapping with custom sort
