+++ packages/lib/treearray/identifier.ts
@@ -1,0 +1,3 @@
+export interface IIdentifier {
+  id: string;
+}
