+++ packages/lib/treearray/package.json
@@ -10,9 +10,9 @@
         "prepack": "yarn test && tsc"
     },
     "devDependencies": {
-        "@types/jest": "^26.0.24",
-        "jest": "^27.0.6",
-        "ts-jest": "^26.4.2",
+        "@types/jest": "^27.0.1",
+        "jest": "^27.1.0",
+        "ts-jest": "^27.0.5",
         "typescript": "^4.3.5"
     },
     "publishConfig": {
